[TT-17486] refactor: derive probe timeout from check interval; fold tests into existing suites
The probe timeout is now half the health-check interval instead of a
fixed 5s cap, so it scales with operator config and the probe always
reports its own error before the round's barrier expires. The TT-17486
regression tests move from tt17486_repro_test.go into
dashboard_register_test.go and health_check_test.go.
